// applications/webapp/src/app/api/generated/model/ocr-frame-group.ts
/**
 * VidIngest API
 *
 * Contact: dev@tradinglabs.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OcrResultDto } from './ocr-result-dto';


export interface OcrFrameGroup { 
    frameId?: string;
    timestampSeconds?: number;
    framePath?: string;
    lines?: Array<OcrResultDto>;
}