// applications/webapp/src/app/api/generated/model/ocr-result-dto.ts
/**
 * VidIngest API
 *
 * Contact: dev@tradinglabs.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OcrResultDto { 
    id?: string;
    frameId?: string;
    text?: string;
    confidence?: number;
    bbox?: Array<Array<number>>;
    language?: string;
    createdAt?: string;
}