// applications/webapp/src/app/api/generated/model/speaker-dto.ts
/**
 * VidIngest API
 *
 * Contact: dev@tradinglabs.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SpeakerDto { 
    id?: string;
    videoId?: string;
    label?: string;
    displayName?: string;
    segmentCount?: number;
    createdAt?: string;
}